Remove anonymous class
